move stuff outside of args.py
